requirements.txt
Django>=3.2,<4.0
wagtail>=4.1,<4.2
psycopg2-binary
django-countries
django-widget-tweaks
wagtail_grapple